Added slider's positioning logic to the component itself
Before the change, the component was positioned using container's flexbox CSS. As far as I know, flexbox is not meant to be used to position stacked elements. Anyway, stacking still worked in all browsers, except on ios 10. Could probably test more, but limited access to the device prevented me from doing it.

With calculating the top/left offsets when viewport changes, we also decoupled component positioning from the rest of the app.