Changed canvas rendering of the higlighted part of the arc to SVG path rendering
Canvas rendering performed the best in terms of smoothines and FPSs on old Apple devices (ipad 2, iphone 5), but it had one downside, which led me to drop the canvas solution.

In some browsers (mobile FF and Samsung Internet) there was sometimes a slight delay after an user interaction. I am not sure, what was causing the jank. The most sensible reason I could come up with, was that the offloading of the textures (compositor layers) to the GPU was sometimes slower than usual.